Code Book for mytidydata.txt
This is the code book for the Coursera Getting and Cleaning Data
Assignment by Abigail Moore. It describes the tidy data set  in file
mytidydata.txt produced by the code run_analysis.R. The "mytidydata.txt"
file has been uploaded to the Coursera assignment page. This file
contains a descripion of the data and code for opening the file
"mytidydata.txt" in R. Note: this file also opens tidily in excel.