Code Book version 2  by Abigail Moore
This code book corrects some typing errors in version 1.  The code book
describes the dataset tproduced for this assignment in a tidy log/narrow
format and stored in the delimited text file "mytidydata.txt". The code
book also gives a code for opening the tidy data file "mytidydata.txt"
in R. The dataset also opens tidily in excel.